Change integration .py to act as package, add run scripts.
Change the functions in integration.py to behave allow other scripts
to run the monte-carlo integration, permitting easy separation of
the two question sections. Add Single_value.py to calculate an estimate
for a particular number of samples and multi_value.py to caluculate
the values for a range of sample numbers.